chore(earn): apy format refactor
