--- packages/suite/src/components/earn/utils/earnApyUtils.ts
@@ -1,2 +1,0 @@
-export const getApyRate = (apyRate: number): number =>
-    Number.isFinite(apyRate) ? apyRate : Number.NEGATIVE_INFINITY;
