--- packages/suite/src/views/wallet/staking/utils/formatStakeValues.tsx
@@ -1,9 +1,0 @@
-import { Translation } from '@suite/intl';
-
-export const formatApyValue = (apy?: number | null) => {
-    if (apy == null) {
-        return <Translation id="TR_STAKE_N_A" />;
-    }
-
-    return `${apy}`;
-};
